Changes: - Refactored EditableFormComponent - Added skeleton for forms
