+++ app/components/Forms/FormComponentBG.tsx
@@ -1,18 +1,23 @@
 "use client";
 
 const FormComponentBG = ({
-  children,
+  children = null,
   className,
   onClick,
+  skeleton = false,
 }: {
-  children: React.ReactNode;
+  children?: React.ReactNode;
   className?: string;
   onClick?: () => void;
+  skeleton?: boolean;
 }) => {
   return (
     <div
-      className={`flex flex-col bg-base-100 shadow-br rounded-lg p-4 py-2 ${className}`}
+      className={`${className} flex flex-col shadow-br rounded-lg p-4 py-2 ${
+        skeleton ? "skeleton h-20" : "bg-base-100"
+      }`}
       onClick={onClick}
+      style={skeleton ? {} : undefined}
     >
       {children}
     </div>
