+++ app/globals.css
@@ -447,3 +447,22 @@ body {
       var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow);
   }
 }
+
+.skeleton-forms-component {
+  border-radius: var(--radius-box) /* var(--radius-box) */;
+  background-color: var(--color-base-100) /* var(--color-base-100) */;
+  @media (prefers-reduced-motion: reduce) {
+    transition-duration: 15s;
+  }
+  will-change: background-position;
+  animation: skeleton 1.8s ease-in-out infinite;
+  background-image: linear-gradient(
+    105deg,
+    #0000 0% 40%,
+    var(--color-base-200) /* var(--color-base-100) */ 50%,
+    #0000 60% 100%
+  );
+  background-size: 200% auto;
+  background-repeat: no-repeat;
+  background-position-x: -50%;
+}
