+++ app/pages/Forms/FormsLoading.tsx
@@ -1,0 +1,22 @@
+import EditableFormHeader from "@/app/components/Forms/EditableFormHeader";
+import FormBG from "@/app/components/Forms/FormBG";
+import FormComponentBG from "@/app/components/Forms/FormComponentBG";
+
+const FormsLoading = () => {
+  return (
+    <>
+      <EditableFormHeader />
+      <FormBG>
+        <FormComponentBG className="h-30" skeleton />
+        <FormComponentBG skeleton />
+        <FormComponentBG className="h-40" skeleton />
+        <FormComponentBG skeleton />
+        <FormComponentBG className="h-50" skeleton />
+        <FormComponentBG skeleton />
+        <FormComponentBG skeleton />
+      </FormBG>
+    </>
+  );
+};
+
+export default FormsLoading;
